패킷 구현중

+++ Server/DBServerASPdotNet/DBServer/DB/MongoDBLib.cs
@@ -52,7 +52,7 @@ namespace DBServer.DB
         public class DBBasicGameData
         {
             public string _id;//user ID
-            public float Rating;//Rating For Match making
+            public int Rating;//Rating For Match making
             public int WinNumber;
             public int LoseNumber;
             //public int Ranking;
